WIP: introduce WeakFixupFn, start removing  Option<*const ()>
This introduce a breaking change starting to modifying all methods that
indicate a realloc by returning a Option(*const()) into stdlib compatible
signatures and variants taking a closure for the fixup.

This commits starts with reserve and shrink methods, more in the following commits.

The `WeakFixupFn` is required for Drain and other iterators which do significant work
in the destructor (possibly reallocating the headervec there). The closure
can be passed along and called when required.

Compatibility note:
When it is not trivially possible to refactor fixup functionality to a closure then
old code can be migrated by:

  let maybe_realloc: Option<*const ()> = {
      let mut opt_ptr = None;
      hv.reserve_with_weakfix(&mut self, 10000, |ptr| opt_ptr = Some(ptr));
      opt_ptr
  };

  // now maybe_realloc is like the old Option<*const ()> return

  if let Some(ptr) = maybe_realloc {
        // fixup code using ptr
  }

Note 2:
do we want legacy wrapper functions that have the old behavior eg.

  #[deprecated("upgrade to the new API")]
  pub fn reserve_legacy(&mut self, additional: usize) -> Option<*const ()> {
      let mut opt_ptr = None;
      self.reserve_with_weakfix(additional, |ptr| opt_ptr = Some(ptr));
      opt_ptr
  };