ADD: a lot `impl From`'s
These are mostly the std Vec compatible From impls. By default this creates `HeaderVec<(), T>`
Additionally when one wants to pass a header one can do that by a `WithHeader(H,T)` tuple
struct. The later is mostly for convenience.

I took my liberty to introduce my xmacro crate here. The macro expands to ~120 lines of code.
When this dependency is not wanted it could be replaced with the expanded code.