assert that T is not a ZST
HeaderVec does not support ZST's, this will give a compiletime error when one tries to do so.
Before it was a Divide by Zero runtime error in the offset calculation.

Remove the ZST code from Drain too (was some remains from std Drain)

Its unlikely that supporting ZST's in HeaderVec makes any sense. Forbiding this altogether
make this clear. When someone comes up with a real use case this may be reviewed.