splice() method and Splice iterator
This implements splicing. This allows for insert/replace operations in HeaderVecs.

The semantic is the same as the stdlib splice. But we use a slightly different algorithm
internally:
* The stdlib Splice collects excess replace_with elements in a temporary vector and later
  fills this in. in some/worst cases this may cause the tail to be moved arouind two times.
* Our implementation pushes objects from `replace_with` directly onto the source HeaderVec
  If tail elements would be overwritten they are stashed on a temporary vector.
  Only these stashed elements are moved twice, the remainder of the tail is moved only once.

Note: moving elements because of reallocation is not accounted here. In both implementations
this can happen equally bad. The stdlib can mitigate that by this by some unstable features.
We try hard to do it as best as possible.

Benchmarks will follow.